v32 Phase A Step 2 — examples の u64 eid 対応
- examples 7 本のキャスト修正(as u32/u64, Vec<u32>→Vec<u64>)
- agentic_workload_bench: query/pull_raw 値を u32 キャスト
- v27_vs_v31_full: subset スライス型、bfs 起点を u64 化
- v27_vs_v31_full: parent tie を tie_async with cast
- delta_sync_experiment: db.get/db.tie の eid を u64 化
- v27_concurrent_bench: eid 計算を u64 化

+++ examples/agentic_workload_bench.rs
@@ -71,7 +71,7 @@ fn main() {
     let t0 = Instant::now();
     let mut hits = 0;
     for _ in 0..QUERY_ITERS {
-        let eids = eng.query(&[("speaker", target_speaker), ("topic", target_topic)]);
+        let eids = eng.query(&[("speaker", target_speaker as u32), ("topic", target_topic as u32)]);
         hits = eids.len();
     }
     let enchu_q1 = t0.elapsed();
@@ -81,7 +81,7 @@ fn main() {
     // Query 2: query + decision text 全件抽出
     let t0 = Instant::now();
     for _ in 0..QUERY_ITERS {
-        let eids = eng.query(&[("speaker", target_speaker), ("topic", target_topic)]);
+        let eids = eng.query(&[("speaker", target_speaker as u32), ("topic", target_topic as u32)]);
         for &e in &eids {
             let _ = eng.get_text(e, "decision");
         }
@@ -94,7 +94,7 @@ fn main() {
     let t0 = Instant::now();
     let mut rev_hits = 0;
     for _ in 0..QUERY_ITERS {
-        let sessions = eng.pull_raw("topic", target_topic);
+        let sessions = eng.pull_raw("topic", target_topic as u32);
         rev_hits = sessions.len();
     }
     let enchu_q3 = t0.elapsed();
